infinite scroll, mongodb and token implementation

+++ staff/ignacio-nuñez/custom-app/src/logic/delete-post.js
@@ -1,11 +1,18 @@
-function deletePost(postId, postUserId, userId) {
+import extractSubFromToken from '../utils/extractSubFromToken'
+
+function deletePost(postId, postUserId, token) {
+    if(!token.length) throw new Error('token is empty')
+    if(typeof token !== 'string') throw new TypeError('token is not a string')
+
+    const userId = extractSubFromToken(token)
+
     if (typeof postId !== 'string') throw new Error('postId is not a string')
     if(postUserId !== userId) throw new Error('userId is different than postUserId')
 
     return fetch(`http://localhost:80/posts/${postId}`, {
         method: 'DELETE',
         headers: { 
-            'Authorization': `Bearer ${userId}`,
+            'Authorization': `Bearer ${token}`,
             'Content-Type': 'application/json'
          }
     })
